+++ src/schemathesis/transports/asgi.py
@@ -1,0 +1,7 @@
+from inspect import iscoroutinefunction
+
+
+def is_asgi_app(app: object) -> bool:
+    return iscoroutinefunction(app) or (
+        hasattr(app, "__call__") and iscoroutinefunction(app.__call__)  # noqa: B004
+    )